NOISSUE - Refactor SDK and things service (#420)
* Refactor Mainflux go SDK

Add structures instead of string parameters. Add offset and limit
parameters to things and channels methods. Add better configuration
support.


* Add new public errors with better error handling


* Update SDK to use uint instread of string id


* Update cli to use new SDK API


* Remove TLS termination from nginx configuration


* Update SDK documentation and structures


* Refactor things service

Decouple HTTP layer from business logic. Remove ID number validation
check. Remove models from HTTP requests and responses.


* Reformat tests for things service


* Increase test coverage for things service

+++ things/api/logging.go
@@ -68,7 +68,7 @@ func (lm *loggingMiddleware) ViewThing(key string, id uint64) (thing things.Thin
 	return lm.svc.ViewThing(key, id)
 }
 
-func (lm *loggingMiddleware) ListThings(key string, offset, limit int) (things []things.Thing, err error) {
+func (lm *loggingMiddleware) ListThings(key string, offset, limit uint64) (things []things.Thing, err error) {
 	defer func(begin time.Time) {
 		message := fmt.Sprintf("Method list_things for key %s took %s to complete", key, time.Since(begin))
 		if err != nil {
@@ -133,7 +133,7 @@ func (lm *loggingMiddleware) ViewChannel(key string, id uint64) (channel things.
 	return lm.svc.ViewChannel(key, id)
 }
 
-func (lm *loggingMiddleware) ListChannels(key string, offset, limit int) (channels []things.Channel, err error) {
+func (lm *loggingMiddleware) ListChannels(key string, offset, limit uint64) (channels []things.Channel, err error) {
 	defer func(begin time.Time) {
 		message := fmt.Sprintf("Method list_channels for key %s took %s to complete", key, time.Since(begin))
 		if err != nil {
